+++ packages/terra-table/src/Table.jsx
@@ -30,6 +30,7 @@ const RowSelectionModes = {
 
 const TableConstants = {
   ROW_SELECTION_COLUMN_WIDTH: 40,
+  TABLE_MARGIN_RIGHT: 15,
 };
 
 const ROW_SELECTION_COLUMN_ID = 'table-rowSelectionColumn';
@@ -540,6 +541,12 @@ function Table(props) {
     }
   };
 
+  // Added margin to allow for resizing of last column.
+  const hasResizableCol = tableColumns[tableColumns.length - 1].isResizable;
+  const tableStyle = {
+    marginRight: hasResizableCol ? `${TableConstants.TABLE_MARGIN_RIGHT}px` : '0',
+  };
+
   // -------------------------------------
 
   return (
@@ -557,6 +564,7 @@ function Table(props) {
         aria-labelledby={ariaLabelledBy}
         aria-label={ariaLabel}
         aria-rowcount={tableRowCount}
+        style={tableStyle} // eslint-disable-line react/forbid-dom-props
         className={cx('table', { headerless: !hasVisibleColumnHeaders })}
         onKeyDown={!isGridContext ? onKeyDown : undefined}
         {...(activeIndex != null && { onMouseUp, onMouseMove, onMouseLeave: onMouseUp })}