+++ packages/terra-data-grid/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added
   * Added rowMinimumHeight prop that gets passed to terra-table to support flexible row height.
 
+* Changed
+  * Updated jest snapshots.
+
 ## 1.11.0 - (January 12, 2024)
 
 * Added