+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Changed
+  * Updated the `terra-date-time-picker` example for Error label.
+
+* Added
+  * Added documentation for exported constants of `terra-table` in the About page.
+  * Added tests for resizable `terra-table`.
+
 * Added
   * Added the 'aria-label' attribute to announce the information for the 'Flowsheet-Data Grid'.